+++ src/util/utils.h
@@ -7,5 +7,5 @@ int centerHeight(int max_y, int obj_height);
 int centerWidth(int max_x, int obj_width);
 
 //math utils
-float normalizeAngle(float angle);
-float getShortestAngleDiff(float target, float current);
+double normalizeAngle(double angle);
+double getShortestAngleDiff(double target, double current);
